Shove all Python scripts into one
